+++ src/models/index.js
@@ -1,6 +1,8 @@
 module.exports = {
+  Egg: require("./egg"),
   Generation: require("./generation"),
   GlobalStat: require("./global-stat"),
+  MaxBattle: require("./max-battle"),
   Move: require("./move"),
   Pokemon: require("./pokemon"),
   PokemonAsset: require("./pokemon-asset"),